switch github button to css module
