+++ src/css/custom.scss
@@ -39,6 +39,10 @@
   // Theme Colors
   // ==============================================================================
 
+  // Primary
+  // See 'General Docusaurs'
+  --ifm-color-on-primary: var(--ifm-color-white);
+
   // Background
   --ifm-color-background: var(--ifm-color-white);
   --ifm-color-background-2: var(--ifm-color-white-dark);
@@ -332,6 +336,10 @@ html[data-theme='dark'] {
   // ==============================================================================
   // Theme Colors
   // ==============================================================================
+
+  // Primary
+  // See 'General Docusaurs'
+  --ifm-color-on-primary: var(--ifm-color-black);
 
   // Background
   --ifm-color-background: var(--ifm-color-black);
