+++ src/pages/LandingPage/index.tsx
@@ -6,8 +6,6 @@ import Layout from "@theme/Layout";
 import "react-toastify/dist/ReactToastify.css";
 import HeaderView from "./components/HeaderView";
 import useDocusaurusContext from "@docusaurus/useDocusaurusContext";
-import GiveItATryView from "./components/GiveItATriyView";
-import Spacer from "../../components/Spacer";
 
 const LandingPage: React.FC = () => {
   const windowSize = useWindowSize();
